refactor(evfs): split tests into modules and add metadata tests
